* refactor: moved page constants to packages
