+++ web/core/components/pages/list/order-by.tsx
@@ -5,8 +5,6 @@ import { ArrowDownWideNarrow, ArrowUpWideNarrow, Check, ChevronDown } from "luci
 import { TPageFiltersSortBy, TPageFiltersSortKey } from "@plane/types";
 // ui
 import { CustomMenu, getButtonStyling } from "@plane/ui";
-// constants
-import { PAGE_SORTING_KEY_OPTIONS } from "@/constants/page";
 // helpers
 import { cn } from "@/helpers/common.helper";
 
@@ -15,6 +13,15 @@ type Props = {
   sortBy: TPageFiltersSortBy;
   sortKey: TPageFiltersSortKey;
 };
+
+const PAGE_SORTING_KEY_OPTIONS: {
+  key: TPageFiltersSortKey;
+  label: string;
+}[] = [
+  { key: "name", label: "Name" },
+  { key: "created_at", label: "Date created" },
+  { key: "updated_at", label: "Date modified" },
+];
 
 export const PageOrderByDropdown: React.FC<Props> = (props) => {
   const { onChange, sortBy, sortKey } = props;
