+++ web/core/components/pages/modals/create-page-modal.tsx
@@ -1,5 +1,6 @@
 import { FC, useEffect, useState } from "react";
-
+// constants
+import { EPageAccess } from "@plane/constants";
 // types
 import { TPage } from "@plane/types";
 // ui
@@ -8,7 +9,6 @@ import { EModalPosition, EModalWidth, ModalCore } from "@plane/ui";
 import { PageForm } from "@/components/pages";
 // constants
 import { PAGE_CREATED } from "@/constants/event-tracker";
-import { EPageAccess } from "@/constants/page";
 // hooks
 import { useProjectPages, useEventTracker } from "@/hooks/store";
 import { useAppRouter } from "@/hooks/use-app-router";
