+++ web/core/store/base-command-palette.store.ts
@@ -1,8 +1,11 @@
 import { observable, action, makeObservable } from "mobx";
-import { EIssuesStoreType, TCreateModalStoreTypes } from "@plane/constants";
-// services
-// types / constants
-import { DEFAULT_CREATE_PAGE_MODAL_DATA, EPageAccess, TCreatePageModal } from "@/constants/page";
+import {
+  EIssuesStoreType,
+  TCreateModalStoreTypes,
+  DEFAULT_CREATE_PAGE_MODAL_DATA,
+  EPageAccess,
+  TCreatePageModal,
+} from "@plane/constants";
 
 export interface ModalData {
   store: EIssuesStoreType;
